Move user selection getting api function to core and call it in github
